feat: refactor question generation logic and move to dedicated hook for improved organization and maintainability
